+++ crates/nostr-database/src/lib.rs
@@ -22,9 +22,6 @@ pub mod flatbuffers;
 pub mod index;
 pub mod memory;
 pub mod profile;
-mod tag_indexes;
-#[cfg(feature = "flatbuf")]
-mod temp;
 
 pub use self::error::DatabaseError;
 #[cfg(feature = "flatbuf")]
@@ -32,8 +29,6 @@ pub use self::flatbuffers::{FlatBufferBuilder, FlatBufferDecode, FlatBufferEncod
 pub use self::index::{DatabaseIndexes, EventIndexResult};
 pub use self::memory::{MemoryDatabase, MemoryDatabaseOptions};
 pub use self::profile::Profile;
-#[cfg(feature = "flatbuf")]
-pub use self::temp::TempEvent;
 
 /// Backend
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -160,13 +155,6 @@ pub trait NostrDatabase: AsyncTraitDeps {
 
     /// Query store with filters
     async fn query(&self, filters: Vec<Filter>, order: Order) -> Result<Vec<Event>, Self::Err>;
-
-    /// Get event IDs by filters
-    async fn event_ids_by_filters(
-        &self,
-        filters: Vec<Filter>,
-        order: Order,
-    ) -> Result<Vec<EventId>, Self::Err>;
 
     /// Get `negentropy` items
     async fn negentropy_items(
@@ -346,17 +334,6 @@ impl<T: NostrDatabase> NostrDatabase for EraseNostrDatabaseError<T> {
 
     async fn query(&self, filters: Vec<Filter>, order: Order) -> Result<Vec<Event>, Self::Err> {
         self.0.query(filters, order).await.map_err(Into::into)
-    }
-
-    async fn event_ids_by_filters(
-        &self,
-        filters: Vec<Filter>,
-        order: Order,
-    ) -> Result<Vec<EventId>, Self::Err> {
-        self.0
-            .event_ids_by_filters(filters, order)
-            .await
-            .map_err(Into::into)
     }
 
     async fn negentropy_items(